+++ @pew/pew-update-produto.php
@@ -1,5 +1,5 @@
 <?php
-    $post_fields = array("id_produto", "sku", "nome", "marca", "preco", "estoque", "estoque_baixo", "tempo_fabricacao", "descricao_curta", "descricao_longa", "peso", "comprimento", "largura", "altura", "status");
+    $post_fields = array("id_produto", "sku", "nome", "marca", "id_cor", "preco", "preco_promocao", "promocao_ativa", "estoque", "estoque_baixo", "tempo_fabricacao", "descricao_curta", "descricao_longa", "url_video", "peso", "comprimento", "largura", "altura", "status");
     $file_fields = array();
     $invalid_fields = array();
     $gravar = true;
@@ -22,16 +22,20 @@
     }
     if($gravar){
         require_once "pew-system-config.php";
+        require_once "@classe-system-functions.php";
+        
         $dataAtual = date("Y-m-d h:i:s");
         /*POST DATA*/
         $idProduto = addslashes($_POST["id_produto"]);
         $skuProduto = addslashes($_POST["sku"]);
         $nomeProduto = addslashes($_POST["nome"]);
         $marcaProduto = addslashes($_POST["marca"]);
+        $idCor = (int)$_POST["id_cor"];
         $precoProduto = $_POST["preco"];
-        $precoPromocaProduto = $_POST["precoPromocao"];
-        $precoProduto = pew_number_format($precoProduto);
-        $precoPromocaProduto = pew_number_format($precoPromocaProduto);
+        $precoProduto = $pew_functions->custom_number_format($precoProduto);
+        $precoPromocaoProduto = $_POST["preco_promocao"];
+        $precoPromocaoProduto = $pew_functions->custom_number_format($precoPromocaoProduto);
+        $promocaoAtiva = intval($_POST["promocao_ativa"]) == 1 ? 1 : 0;
         $estoqueProduto = (int)$_POST["estoque"] != "" ? (int)$_POST["estoque"] : 0;
         $estoqueBaixoProduto = (int)$_POST["estoque_baixo"] != "" ? (int)$_POST["estoque_baixo"] : 1;
         $tempoFabricacaoProduto = (int)$_POST["tempo_fabricacao"] != "" ? (int)$_POST["tempo_fabricacao"] : 1;;
@@ -41,47 +45,57 @@
         $comprimentoProduto = floatval($_POST["comprimento"]);
         $larguraProduto = floatval($_POST["largura"]);
         $alturaProduto = floatval($_POST["altura"]);
+        $departamentosProduto = isset($_POST["departamentos"]) ? $_POST["departamentos"] : "";
         $categoriasProduto = isset($_POST["categorias"]) ? $_POST["categorias"] : "";
         $especificacoes = isset($_POST["especicacao_produto"]) ? $_POST["especicacao_produto"] : "";
         $produtosRelacionados = isset($_POST["produtos_relacionados"]) ? $_POST["produtos_relacionados"] : "";
+        $coresRelacionadas = isset($_POST["cores_relacionadas"]) ? $_POST["cores_relacionadas"] : "";
         $subcategoriasProduto = isset($_POST["subcategorias"]) ? $_POST["subcategorias"] : "";
         $statusProduto = intval($_POST["status"]) == 1 ? 1 : 0;
-        $statusPromocaoProduto = intval($_POST["statusPromocao"]) == 1 ? 1 : 0;
+        $urlVideoProduto = addslashes($_POST["url_video"]);
+        
+        $http = substr($urlVideoProduto, 0, 5);
+        if($http != "http:" && $http != "https" && $urlVideoProduto != ""){
+            $urlVideoProduto = "http://".$urlVideoProduto;
+        }
+        
         /*END POST DATA*/
         
+
         /*DIR VARS*/
         $dirImagensProdutos = "../imagens/produtos/";
         /*END DIR VARS*/
-        
+
         /*SET TABLES*/
         $tabela_produtos = $pew_custom_db->tabela_produtos;
         $tabela_imagens = $pew_custom_db->tabela_imagens_produtos;
+        $tabela_cores = $pew_custom_db->tabela_cores;
         $tabela_categorias = $pew_db->tabela_categorias;
         $tabela_subcategorias = $pew_db->tabela_subcategorias;
+        $tabela_departamentos = $pew_custom_db->tabela_departamentos;
+        $tabela_departamentos_produtos = $pew_custom_db->tabela_departamentos_produtos;
         $tabela_categorias_produtos = $pew_custom_db->tabela_categorias_produtos;
         $tabela_subcategorias_produtos = $pew_custom_db->tabela_subcategorias_produtos;
         $tabela_produtos_relacionados = $pew_custom_db->tabela_produtos_relacionados;
+        $tabela_cores_relacionadas = $pew_custom_db->tabela_cores_relacionadas;
         $tabela_especificacoes_produtos = $pew_custom_db->tabela_especificacoes_produtos;
         /*END SET TABLES*/
-        
-        if(!function_exists("stringFormat")){
-            function stringFormat($string){
-                $string = str_replace("Ç", "c", $string);
-                $string = str_replace("ç", "c", $string);
-                $string = preg_replace(array("/(á|à|ã|â|ä)/","/(Á|À|Ã|Â|Ä)/","/(é|è|ê|ë)/","/(É|È|Ê|Ë)/","/(í|ì|î|ï)/","/(Í|Ì|Î|Ï)/","/(ó|ò|õ|ô|ö)/","/(Ó|Ò|Õ|Ô|Ö)/","/(ú|ù|û|ü)/","/(Ú|Ù|Û|Ü)/","/(ñ)/","/(Ñ)/"),explode(" ","a A e E i I o O u U n N"), $string);
-                $string = strtolower($string);
-                $string = str_replace("/", "-", $string);
-                $string = str_replace("|", "-", $string);
-                $string = str_replace(" ", "-", $string);
-                $string = str_replace(",", "", $string);
-                return $string;
-            }
-        }
-        
+
         if($nomeProduto != ""){
             echo "<h3 align=center>Gravando dados...</h3>";
-            mysqli_query($conexao, "update $tabela_produtos set sku = '$skuProduto', nome = '$nomeProduto', marca = '$marcaProduto', preco = '$precoProduto', preco_promocao = '$precoPromocaProduto', promocao_ativa = '$statusPromocaoProduto', estoque = '$estoqueProduto', estoque_baixo = '$estoqueBaixoProduto', tempo_fabricacao = '$tempoFabricacaoProduto', descricao_curta = '$descricaoCurtaProduto', descricao_longa = '$descricaoLongaProduto', peso = '$pesoProduto', comprimento = '$comprimentoProduto', largura = '$larguraProduto', altura = '$alturaProduto', data = '$dataAtual', status = '$statusProduto' where id = '$idProduto'");
             
+            $condicaoCor = "id = '$idCor'";
+            $idCor = $pew_functions->contar_resultados($tabela_cores, $condicaoCor) > 0 ? $idCor : null;
+            
+            mysqli_query($conexao, "update $tabela_produtos set sku = '$skuProduto', nome = '$nomeProduto', marca = '$marcaProduto', id_cor = '$idCor', preco = '$precoProduto', preco_promocao = '$precoPromocaoProduto', promocao_ativa = '$promocaoAtiva', estoque = '$estoqueProduto', estoque_baixo = '$estoqueBaixoProduto', tempo_fabricacao = '$tempoFabricacaoProduto', descricao_curta = '$descricaoCurtaProduto', descricao_longa = '$descricaoLongaProduto', url_video = '$urlVideoProduto', peso = '$pesoProduto', comprimento = '$comprimentoProduto', largura = '$larguraProduto', altura = '$alturaProduto', data = '$dataAtual', status = '$statusProduto' where id = '$idProduto'");
+
+            /*ATUALIZA DEPARTAMENTOS*/
+            if($departamentosProduto != ""){
+                mysqli_query($conexao, "delete from $tabela_departamentos_produtos where id_produto = '$idProduto'");
+                foreach($departamentosProduto as $idDepartamento){
+                    mysqli_query($conexao, "insert into $tabela_departamentos_produtos (id_produto, id_departamento) values ('$idProduto', '$idDepartamento')");
+                }
+            }
             /*ATUALIZA CATEGORIAS DO PRODUTO*/
             if($categoriasProduto != ""){
                 $queryCategoriasAtuais = mysqli_query($conexao, "select id_categoria from $tabela_categorias_produtos where id_produto = '$idProduto'");
@@ -111,7 +125,7 @@
                 mysqli_query($conexao, "delete from $tabela_categorias_produtos where id_produto = '$idProduto'");
             }
             /*FIM ATUALIZA CATEGORIAS DO PRODUTO*/
-            
+
             /*ATUALIZA SUBCATEGORIAS DO PRODUTO*/
             if($subcategoriasProduto != ""){
                 $querySubcategoriasAtuais = mysqli_query($conexao, "select id_subcategoria, titulo_subcategoria from $tabela_subcategorias_produtos where id_produto = '$idProduto'");
@@ -147,44 +161,49 @@
                 mysqli_query($conexao, "delete from $tabela_subcategorias_produtos where id_produto = '$idProduto'");
             }
             /*FIM ATUALIZA SUBCATEGORIAS DO PRODUTO*/
-            
+
             /*ATUALIZA IMAGENS DO PRODUTO*/
             $maxImagens = isset($_POST["maximo_imagens"]) && (int)$_POST["maximo_imagens"] ? (int)$_POST["maximo_imagens"] : 4;
-            for($i = 0; $i <= $maxImagens; $i++){
-                $posicaoAnterior = $i - 1;
-                $contarImagem = mysqli_query($conexao, "select count(id) as total_imagem from $tabela_imagens where posicao = '$posicaoAnterior' and id_produto = '$idProduto'");
-                $contagem = mysqli_fetch_assoc($contarImagem);
-                if($contagem["total_imagem"] == 0 && $posicaoAnterior > 0){
+            for($i = 1; $i <= $maxImagens; $i++){
+                $posicaoAnterior = $i - 1;                
+                $condicaoImagem = "id_produto = '$idProduto' and posicao = '$posicaoAnterior'";
+                $totalImagem = $pew_functions->contar_resultados($tabela_imagens, $condicaoImagem);
+                
+                if($totalImagem == 0 && $posicaoAnterior > 0){
                     $posicao = $posicaoAnterior;
                 }else{
                     $posicao = $i;
                 }
+                
                 if(isset($_FILES["imagem$i"])){
-                    $contarImagem = mysqli_query($conexao, "select count(id) as total_imagem from $tabela_imagens where id_produto = '$idProduto' and posicao = '$posicao'");
-                    $contagem = mysqli_fetch_assoc($contarImagem);
-                    $totalImgPosicao = $contagem["total_imagem"];
+                    $condicaoPosImagem = "id_produto = '$idProduto' and posicao = '$posicao'";
+                    $totalImgPosicao = $pew_functions->contar_resultados($tabela_imagens, $condicaoPosImagem);
                     $nomeIMG = $_FILES["imagem$i"]["name"];
                     if($nomeIMG != ""){
-                        $ref = substr(md5(uniqid(time())), 0, 4);
-                        $ref = $ref.$posicao;
-                        $nomeFoto = stringFormat($nome);
                         $ext = pathinfo($_FILES["imagem$i"]["name"], PATHINFO_EXTENSION);
-                        $nomeFoto = $nomeFoto."-product-image-".$ref.".".$ext;
-                        move_uploaded_file($_FILES["imagem$i"]["tmp_name"], $dirImagensProdutos.$nomeFoto);
+                        $ref = substr(md5($nomeProduto.$posicao), 0, 4);
+                        $urlTitulo = $pew_functions->url_format($nomeProduto);
+                        $nomeFinalImagem = $urlTitulo."-".$ref.".".$ext;
+                        move_uploaded_file($_FILES["imagem$i"]["tmp_name"], $dirImagensProdutos.$nomeFinalImagem);
+                        
                         if($totalImgPosicao > 0){
                             $queryImagem = mysqli_query($conexao, "select imagem from $tabela_imagens where id_produto = '$idProduto' and posicao = '$posicao'");
                             $arrayImagem = mysqli_fetch_array($queryImagem);
                             $imagem = $arrayImagem["imagem"];
-                            unlink($dirImagensProdutos.$imagem);
-                            mysqli_query($conexao, "update $tabela_imagens set imagem = '$nomeFoto', status = 1 where id_produto = '$idProduto' and posicao = '$posicao'");
+                            
+                            if(file_exists($dirImagensProdutos.$imagem) && $imagem != ""){
+                                unlink($dirImagensProdutos.$imagem);
+                            }
+                            
+                            mysqli_query($conexao, "update $tabela_imagens set imagem = '$nomeFinalImagems', status = 1 where id_produto = '$idProduto' and posicao = '$posicao'");
                         }else{
-                            mysqli_query($conexao, "insert into $tabela_imagens (id_produto, imagem, posicao, status) values ('$idProduto', '$nomeFoto', '$posicao', 1)");
+                            mysqli_query($conexao, "insert into $tabela_imagens (id_produto, imagem, posicao, status) values ('$idProduto', '$nomeFinalImagem', '$posicao', 1)");
                         }
                     }
                 }
             }
             /*FIM ATUALIZA IMAGENS DO PRODUTO*/
-            
+
             /*INSERE ESPECIFICAÇÕES PRODUTO*/
             if($especificacoes != ""){
                 mysqli_query($conexao, "delete from $tabela_especificacoes_produtos where id_produto = '$idProduto'");
@@ -195,7 +214,7 @@
                     mysqli_query($conexao, "insert into $tabela_especificacoes_produtos (id_especificacao, id_produto, descricao) values ('$idEspecificacao', '$idProduto', '$descricaoEspecificacao')");
                 }
             }
-            
+
             /*ATUALIZA PRODUTOS RELACIONADOS*/
             if($produtosRelacionados != ""){
                 $queryProdutosRelacionados = mysqli_query($conexao, "select * from $tabela_produtos_relacionados where id_produto = '$idProduto' group by id_relacionado");
@@ -225,6 +244,29 @@
                 mysqli_query($conexao, "delete from $tabela_produtos_relacionados where id_produto = '$idProduto'");
             }
             /*FIM ATUALIZA PRODUTOS RELACIONADOS*/
+            
+            /*ATUALIZA CORES DE PRODUTOS RELACIONADOS*/
+            
+            $totalCoresRelacionadas = $pew_functions->contar_resultados($tabela_cores_relacionadas, "id_produto = '$idProduto'");
+            if($totalCoresRelacionadas > 0){
+                $queryRelacionado = mysqli_query($conexao, "select id_relacao from $tabela_cores_relacionadas where id_produto = '$idProduto'");
+                while($infoCorRelacionada = mysqli_fetch_array($queryRelacionado)){
+                    $idProdRel = $infoCorRelacionada["id_relacao"];
+                    mysqli_query($conexao, "delete from $tabela_cores_relacionadas where id_produto = '$idProduto' and id_relacao = '$idProdRel'");
+                    mysqli_query($conexao, "delete from $tabela_cores_relacionadas where id_produto = '$idProdRel' and id_relacao = '$idProduto'");
+                }
+                
+            }
+            
+            if($coresRelacionadas != ""){
+                foreach($coresRelacionadas as $idProdutoRelacionado){
+                    mysqli_query($conexao, "insert into $tabela_cores_relacionadas (id_produto, id_relacao, data_controle, status) values ('$idProduto', '$idProdutoRelacionado', '$dataAtual', 1)");
+                    
+                    mysqli_query($conexao, "insert into $tabela_cores_relacionadas (id_produto, id_relacao, data_controle, status) values ('$idProdutoRelacionado', '$idProduto', '$dataAtual', 1)");
+                }
+            }   
+            
+            /*END ATUALIZA CORES DE PRODUTOS RELACIONADOS*/
             
             echo "<script>window.location.href='pew-edita-produto.php?msg=Produto atualizado com sucesso&msgType=success&id_produto=$idProduto';</script>";
         }else{
